Modified pretty extensively to bring up to date and add shallow cloning

+++ build.sh
@@ -1,8 +1,44 @@
 #!/bin/bash
-if [ -z $LIB_BUILDER_PATH ]; then
-       	LIB_BUILDER_PATH=.
+if ! [ -x "$(command -v python)" ]; then
+        echo "ERROR: python is not installed! Please install python first."
+        exit 1
 fi
-source $LIB_BUILDER_PATH/tools/config.sh
+
+if ! [ -x "$(command -v git)" ]; then
+        echo "ERROR: git is not installed! Please install git first."
+        exit 1
+fi
+
+if ! [ -x "$(command -v make)" ]; then
+        echo "ERROR: Make is not installed! Please install Make first."
+        exit 1
+fi
+
+if ! [ -x "$(command -v flex)" ]; then
+        echo "ERROR: flex is not installed! Please install flex first."
+        exit 1
+fi
+
+if ! [ -x "$(command -v bison)" ]; then
+        echo "ERROR: bison is not installed! Please install bison first."
+        exit 1
+fi
+
+if ! [ -x "$(command -v gperf)" ]; then
+        echo "ERROR: gperf is not installed! Please install gperf first."
+        exit 1
+fi
+
+if ! [ -x "$(command -v stat)" ]; then
+        echo "ERROR: stat is not installed! Please install stat first."
+        exit 1
+fi
+
+source $(dirname $0)/tools/config.sh
+
+# clone/update repositories from git
+$AR_ROOT/tools/pull-repos.sh
+if [ $? -ne 0 ]; then exit 1; fi
 
 if [ -e "$AR_TOOLS" ]; then
         rm -rf "$AR_TOOLS"
@@ -11,28 +47,21 @@ if [ -e "$AR_SDK" ]; then
         rm -rf "$AR_SDK"
 fi
 
-#Build esp32
-export PRODUCT=esp32
-cp $LIB_BUILDER_PATH/sdkconfig.$PRODUCT $LIB_BUILDER_PATH/sdkconfig
-$IDF_PATH/tools/idf.py fullclean
-$IDF_PATH/tools/idf.py build
+source $IDF_PATH/export.sh
+
+builder esp32
+$AR_ROOT/tools/prepare-libs.sh
 if [ $? -ne 0 ]; then exit 1; fi
-$LIB_BUILDER_PATH/tools/prepare-libs.sh
-if [ $? -ne 0 ]; then exit 1; fi
-$LIB_BUILDER_PATH/tools/build-bootloaders.sh
+$AR_ROOT/tools/build-bootloaders.sh
 if [ $? -ne 0 ]; then exit 1; fi
 
-#Build esp32s2
-cp $LIB_BUILDER_PATH/sdkconfig.esp32s2 $LIB_BUILDER_PATH/sdkconfig
-$IDF_PATH/tools/idf.py fullclean
-$IDF_PATH/tools/idf.py build
+builder esp32s2
+$AR_ROOT/tools/prepare-libs.sh
 if [ $? -ne 0 ]; then exit 1; fi
-$LIB_BUILDER_PATH/tools/prepare-libs.sh
-if [ $? -ne 0 ]; then exit 1; fi
-$LIB_BUILDER_PATH/tools/build-bootloaders.sh
+$AR_ROOT/tools/build-bootloaders.sh
 if [ $? -ne 0 ]; then exit 1; fi
 
 # archive the build
-$LIB_BUILDER_PATH/tools/archive-build.sh
+$AR_ROOT/tools/archive-build.sh
 if [ $? -ne 0 ]; then exit 1; fi
